requirements.txt
streamlit>=1.20
openai>=0.27.0
google-generativeai>=0.7.2
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0
tqdm
